icometrix/dicom2nifti#135: Make the setting of echo time and repetition time failsafe in case it is empty,null or an invalid input. icometrix/dicom2nifti#136: Add support for SWI containing both phase and magnitude encoding
