+++ dicom2nifti/convert_generic.py
@@ -17,6 +17,124 @@ from dicom2nifti.exceptions import ConversionError, ConversionValidationError
 
 logger = logging.getLogger(__name__)
 
+def get_grouped_dicoms(dicom_input):
+    """
+    Search all dicoms in the dicom directory, sort and validate them
+
+    fast_read = True will only read the headers not the data
+    """
+
+    # Order all dicom files by InstanceNumber
+    dicoms = sorted(dicom_input, key=lambda x: x.InstanceNumber)
+
+    # now group per stack
+    grouped_dicoms = [[]]  # list with first element a list
+    stack_index = 0
+
+    # loop over all sorted dicoms and sort them by stack
+    # for this we use the position and direction of the slices so we can detect a new stack easily
+    previous_position = None
+    previous_direction = None
+    for dicom_ in dicoms:
+        current_direction = None
+        # if the stack number decreases we moved to the next stack
+        if previous_position is not None:
+            current_direction = numpy.array(dicom_.ImagePositionPatient) - previous_position
+            current_direction = current_direction / numpy.linalg.norm(current_direction)
+        if current_direction is not None and \
+                previous_direction is not None and \
+                not numpy.allclose(current_direction, previous_direction, rtol=0.05, atol=0.05):
+            previous_position = numpy.array(dicom_.ImagePositionPatient)
+            previous_direction = None
+            stack_index += 1
+        else:
+            previous_position = numpy.array(dicom_.ImagePositionPatient)
+            previous_direction = current_direction
+
+        if stack_index >= len(grouped_dicoms):
+            grouped_dicoms.append([])
+        grouped_dicoms[stack_index].append(dicom_)
+
+    return grouped_dicoms
+
+def is_4d(grouped_dicoms):
+    """
+    Use this function to detect if a dicom series is a ge 4d dataset
+    NOTE: Only the first slice will be checked so you can only provide an already sorted dicom directory
+    (containing one series)
+    """
+    # read dicom header
+    header = grouped_dicoms[0][0]
+
+    # check if contains multiple stacks
+    if len(grouped_dicoms) > 1:
+        return True
+
+    return False
+
+def four_d_to_nifti(grouped_dicoms, output_file):
+    """
+    This function will convert ge 4d series to a nifti
+    """
+
+    # Create mosaic block
+    logger.info('Creating data block')
+    full_block = _get_full_block(grouped_dicoms)
+
+    logger.info('Creating affine')
+    # Create the nifti header info
+    affine, slice_increment = common.create_affine(grouped_dicoms[0])
+
+    logger.info('Creating nifti')
+    # Convert to nifti
+    nii_image = nibabel.Nifti1Image(full_block, affine)
+    common.set_tr_te(nii_image, grouped_dicoms[0][0].RepetitionTime, grouped_dicoms[0][0].EchoTime)
+    logger.info('Saving nifti to disk %s' % output_file)
+
+    # Save to disk
+    if output_file is not None:
+        nii_image.header.set_slope_inter(1, 0)
+        nii_image.header.set_xyzt_units(2)  # set units for xyz (leave t as unknown)
+        nii_image.to_filename(output_file)
+
+    return {'NII_FILE': output_file,
+            'NII': nii_image,
+            'MAX_SLICE_INCREMENT': slice_increment}
+
+def _get_full_block(grouped_dicoms):
+    """
+    Generate a full datablock containing all timepoints
+    """
+    # For each slice / mosaic create a data volume block
+    data_blocks = []
+    for index in range(0, len(grouped_dicoms)):
+        logger.info('Creating block %s of %s' % (index + 1, len(grouped_dicoms)))
+        data_blocks.append(_timepoint_to_block(grouped_dicoms[index]))
+
+    # Add the data_blocks together to one 4d block
+    size_x = numpy.shape(data_blocks[0])[0]
+    size_y = numpy.shape(data_blocks[0])[1]
+    size_z = numpy.shape(data_blocks[0])[2]
+    size_t = len(data_blocks)
+    full_block = numpy.zeros((size_x, size_y, size_z, size_t), dtype=data_blocks[0].dtype)
+    for index in range(0, size_t):
+        if full_block[:, :, :, index].shape != data_blocks[index].shape:
+            logger.warning('Missing slices (slice count mismatch between timepoint %s and %s)' % (index - 1, index))
+            logger.warning('---------------------------------------------------------')
+            logger.warning(full_block[:, :, :, index].shape)
+            logger.warning(data_blocks[index].shape)
+            logger.warning('---------------------------------------------------------')
+            raise ConversionError("MISSING_DICOM_FILES")
+        full_block[:, :, :, index] = data_blocks[index]
+
+    return full_block
+
+def _timepoint_to_block(timepoint_dicoms):
+    """
+    Convert slices to a block of data by reading the headers and appending
+    """
+    # similar way of getting the block to anatomical however here we are creating the dicom series our selves
+    return common.get_volume_pixeldata(timepoint_dicoms)
 
 def multiframe_to_nifti(dicom_input, output_file):
     """
@@ -68,7 +186,7 @@ def multiframe_to_nifti(dicom_input, output_file):
 
     # Set TR and TE if available
     if Tag(0x0018, 0x0080) in dicom_input[0] and Tag(0x0018, 0x0081) in dicom_input[0]:
-        common.set_tr_te(nii_image, float(dicom_input[0].RepetitionTime), float(dicom_input[0].EchoTime))
+        common.set_tr_te(nii_image, dicom_input[0].RepetitionTime, dicom_input[0].EchoTime)
 
     # Save to disk
     if output_file is not None:
@@ -122,6 +240,13 @@ def dicom_to_nifti(dicom_input, output_file):
         # validate that we have an orthogonal image (to detect gantry tilting etc)
         common.validate_orthogonal(dicom_input)
 
+    logger.info('Reading and sorting dicom files')
+    grouped_dicoms = get_grouped_dicoms(dicom_input)
+
+    if is_4d(grouped_dicoms):
+        logger.info('Found sequence type: 4D')
+        return four_d_to_nifti(grouped_dicoms, output_file)
+
     # sort the dicoms
     dicom_input = common.sort_dicoms(dicom_input)
 
@@ -157,7 +282,7 @@ def dicom_to_nifti(dicom_input, output_file):
 
     # Set TR and TE if available
     if Tag(0x0018, 0x0080) in dicom_input[0] and Tag(0x0018, 0x0081) in dicom_input[0]:
-        common.set_tr_te(nii_image, float(dicom_input[0].RepetitionTime), float(dicom_input[0].EchoTime))
+        common.set_tr_te(nii_image, dicom_input[0].RepetitionTime, dicom_input[0].EchoTime)
 
     # Save to disk
     if output_file is not None:
